HZN-1202: Separated SNMP-specific CollectionAgent implementations
